using long intead of int to avoid integer overflow

+++ src/main/scala/AST.scala
@@ -5,7 +5,7 @@ case class FunctionDeclaration(name:String, params: List[String], body: Expressi
 //expressions
 sealed abstract class Expression
 case class ExpVariable(v:String) extends  Expression
-case class ExpInt(v: Int) extends Expression
+case class ExpInt(v: Long) extends Expression
 case class ExpList(v: List[Expression]) extends Expression
 case class ExpCond(p: Predicate, e1: Expression, e2:Expression)  extends  Expression
 case class ExpTryCatch(tryExp: Expression, handlerList: List[Handler]) extends Expression
@@ -20,7 +20,7 @@ case class Predicate(name: String, params: List[Expression])
 
 //values
 sealed abstract class Value
-case class ValInt(v: Int) extends Value
+case class ValInt(v: Long) extends Value
 case class ValList(v: List[Value]) extends Value
 case class ValUncaughtException(v: String) extends Value
